+++ CHANGELOG.md
@@ -11,3 +11,8 @@
 * Add mute/unmute support.
 * Break apart state into separate notifiers.
 * Update example app to take advantage of both (plus sendText).
+
+## 0.0.4
+
+* Changed implementation of mute/unmute. It's now `micMuted` and `speakerMuted`
+* Added functions for toggling mute of mic (`toggleMicMuted()`) and speaker (`toggleSpeakerMuted()`)